+++ python_plus/python_plus/egg-info/CHANGELOG.rst
@@ -1,3 +1,8 @@
+2.0.10.1 (2023-11-17)
+~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] new python version assignment from odoo version
+
 2.0.10 (2023-07-18)
 ~~~~~~~~~~~~~~~~~~~